CDMS-713: Use the provided proxy if running in CDP
The BrowserStack Local binary is currently failing to download for some reason, possibly proxy related.
I don't think the proxy is being used properly, so this updates the browserstack config file to check if the CDP_HTTP_PROXY variable is set (are we running in CDP) and then configures it to use the provided localhost proxy server.