CDMS-713: Separate capabilities, cleanup and make running locally easier
This separates out the capabilities (browsers tested) from the main wdio config file to make it easier to maintain.  It also adds a few from the gov.uk recommended list.

To make it easier to run the BrowserStack tests both locally and in CDP the files were split and the CDP one inherits the local one.

The GitHub files have been removed since they will not work, due to the inability to access CDP from GitHub.